separate out Kotlin classes
